AIstring: change type of `length` to `cuint`
The `length` was changed to a `uint32` (`cuint` in nim) in a recent version of assimp, which has a different size to `csize_t`. This results in the first 4 characters of the `data` member being tacked onto the `length` member, resulting in an impossibly long number for the length of the data. This results in memory issues, etc.